% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisjoin.R
\name{lisjoin}
\alias{lisjoin}
\title{Join a tibble and a named list}
\usage{
lisjoin(.tibble, .list, .key, name = "val", strategy = "left", types = NULL)
}
\arguments{
\item{.tibble}{A tibble to be joined to \code{.list}}

\item{.list}{a named list. Passing an unnamed list will throw an error.}

\item{name}{name of the new variable with matched values from \code{.list}. Defaults to "val".}

\item{strategy}{How should values be matched? Specify either 'left', 'right', 'inner' or 'full' to use the matching \code{dplyr::x_join}. Defaults to 'left'.}

\item{types}{Should the operation enforce types? Defaults to no types, but users can enforce by choosing among 'int', 'dbl', 'chr' and 'lgl'.}

\item{key}{a symbol containing the variable name in \code{.tibble} to be used as key.}
}
\description{
Join a tibble and a named list
}
